+++ CHANGELOG.md
@@ -37,6 +37,7 @@ A new `nand` module with the `NAND` class is added to read and write to NAND ima
   * Add custom section ID for GodMode9 bonus volume
 * Add documentation files created with sphinx-autodoc
 * Switch Sphinx theme to rtd, add example-cia, add info to index page
+* Move RomFS header loading in `RomFSReader` to a new `RomFSLv3Header` class
 
 ## v0.6.0 - January 26, 2022
 ### Highlights